refactored to remove old (multiple)classifierresultscollection, renamed new file to the current classifierresultscollection, and ported over the basic getinfo from old mcrc
